Update Visual Studio version and refine package structure
Updated the Visual Studio version in `Esc.Sdk.Cli.sln` to 17.0.31717.71. Removed the `_SolutionItems` project section, which included `Esc.nuspec` and `README_esc.md`. Modified the `<files>` section in `Esc.nuspec` to replace wildcard entries for platform-specific binaries with explicit entries for x64 and arm64 architectures, clarifying the file structure and targets for the package.